requirements.txt
facebook-scraper>=0.2.0,<1
python-dotenv>=1.0.0,<2
nextcord>=2.1.0,<3